Centralize logic for revision retrieval for lazy variables
This is to avoid rapid growing of the amount of keys and
code duplication. Move the retrieval procedure ("from page",
"parent of") from method key to the params.

For example, "revision-age-by-id" and "revision-age-by-title"
are merged to "revision-age". The 'recipe' for which revision
should be retrieved is indicated by the params.

Bug: T382794
Change-Id: Id72299c2f8a29f9b9ccf08025cfe59cce4823cc1