AbuseLogPrivateDetails: avoid required parameter with empty default
For required parameters, empty-ish values are considered the same as a
missing parameter. This is also true of the default value, so validation
would fail. This can be seen locally by setting
`$wgAbuseFilterPrivateDetailsForceReason = true` and running
ApiStructureTest.

So, check the value of the config flag and use that to determine whether
the parameter should be required XOR have a default (and never both at
the same time).

Change-Id: Ia25d9c11adf25ee35f798fff9d5171862cbb4eae