AbuseLogger: properly distinguish between global filters and central DB
AbuseLogger deals with two similar and related concepts: global filters,
and logs in a central wiki database. While the concept of "central wiki
database" is pretty straightforward, care needs to be taken with central
logs: because we're dealing with two wikis, locality is swapped: a
filter that is global in the context of the current (non-central) wiki
is going to be local in the context of the central wiki.

In hindsight, `insertGlobalLogEntries` was a bad name: it seems to imply
that it's inserting log entries for global filters, but really what it's
doing is inserting log entries on the central wiki. And this certainly
doesn't help clarify.

The previous code was conflating the two concepts, assuming that all
local entries were for local filters, and all foreign entries for global
filters. So, use separate method parameters to determine whether a
filter is global, and whether the dump should be stored centrally.

Follow-up to Iaa56fe44d44a0c.

Bug: T390904
Change-Id: I9f215e3d7b39f18121783e9e5bcdb93d1234e98d