Declare two hook interfaces that don’t return void
These hooks are run as 'abortable' => false, but (in the interfaces)
don’t return void (though in the hook runner class they do). MediaWiki
will start to complain about this, so register it as an exception in the
hook runner test.

Ideally, this would be fixed at some point, but unfortunately changing
an interface’s return type from absent to void is a breaking change
(implementations that don’t return void will fail as incompatible), so I
don’t know if this is possible without replacing the hook interface with
a new version.

Change-Id: Ibde07eeef76bec9c14f34d8ce99f74bf22108b0d
Needed-By: I858d5dc02c65a0c68081214c4c71a7d816c0a524