refactor: Check whether BlockedDomain should be enabled in ServiceWiring
Previously, AbuseFilter checked whether the BlockedDomain
feature was enabled in BlockedDomainFilter, by getting
the variable in by the global keyword.

This patch moves this check to ServiceWiring, using
$services->getMainConfig() instead. If the feature
is disabled, a special version of the class,
NoopBlockedDomainFilter is returned.

A noop implementation is used, because integrating with
CommunityConfiguration would require not loading
the actual class at all if CC is not enabled.

Bug: T393240
Change-Id: I1f266e15392adc274e7c1eaede0d61dbd7bfd00e