AbuseFilter: Add support for suppressing filters
Adds a "suppressed" flag for filters that may contain sensitive
information (e.g., PII). Suppression restricts visibility of the
filter's details, logs, and history to users with the
'viewsuppressed' right. All logs from a suppressed filter are
effectively auto-suppressed.

Unlike the "protected" flag, suppression does not require certain
variables to be used in the filter.

* Add a "suppressed" checkbox in the filter edit form, usable only by
  users with the 'suppressrevision' right.
* Restrict suppressed filter details, history, and log
  visibility to users with the 'viewsuppressed' right.
* Add support for showing and managing suppressed filters through the
  UI and API.
* Add unit and integration tests for suppressed filters.

Suppression operates independently from making a filter private.

Therefore, a suppressed filter can be viewed by any user with the
'viewsuppressed' right. If the filter is also private, a user must
have both 'viewsuppressed' and 'abusefilter-view-private' rights to
view it.

Bug: T290324
Change-Id: Ibd59bffb481426a5ba6ad0bb74e74d400c5ad47b