Refactor Filters to use UserIdentity
Filters provide methods to get/set userId and userName. Instead of
handling userId and userName seperately we can use the UserIdentity
istead and pass it as a single object. This simplifies the interface
as there is no need for setters/getters for both and prevents
from situations where only one setter is called.

Additionally this change will be required to:
 - refactor call to deprecated Linker::userLink() as the replacement
 requires a UserIdentity as input instead of id and name pair
 - refactor call to trigger an Echo event, as the RecipientSet
 requires recipients in form of UserIdentity[]

Change-Id: I19e6a3a3d81fb2b130fc2c8d0e51465784987073