EchoNotifierTest should test Event, not internal getDataForEvent
When performing an integration test, the EchoNotifier should test
Echo as a blackbox, not test only that some method returns a specific
array.
This change allows us to make getDataForEvent private as previously
it had to be exposed to allow testing. This also will allow us
to pass Recipients from notifyForFilter() without having to call
getFilterObject() twice which may be expensive and at the end redundant.

Change-Id: Ic204a876adbd40a587899d9a32cc415688e8783c