Pass RecipientSet instead of using Event::RECIPIENTS_IDX
Recommended way to pass recipients with Echo Event is to use the
second optional argument $recipientSet. Previously we could pass
the recipients via $info['extra'] array. To simplify the code a bit
and unify the Echo interface with recent NotifyService, all callers
should follow the new interface.

Passing recipients via extra array is still supported as this is how
recipients information is stored internally in Echo, altough it is not
recommended. Passing set of UserIdentities gives a better control and
type safety.

Change-Id: Ie722558c8cd068210992bded9cb13cde768c0c86