fix: Do not fail if CommunityConfiguration is not installed
Follow-up to I9fd092ce0a4a098dd10f8bf6acadc73708031ba9.

The patch started to inject CommunityConfiguration.ProviderFactory
to EditPermissionHandler. This was left-over from a previous version
of the patch, and I forgot to remove that line.

AbuseFilter is meant to work even if CommunityConfiguration is missing,
and this injection breaks that. Since it is not needed, just remove it.

In addition to breaking expectations, this also breaks WMF production,
as loginwiki has AbuseFilter, but does not have CommunityConfiguration.

Bug: T401843
Change-Id: I45311970e5da673e5d27d0445dd94c9387820c63