Set proper keygroups in BagOStuff makeKey calls
There were numerous keys representing different computations from
different parts of the codebase, all under a single "abusefilter"
keygroup. This creates makes stats much less useful [1].

Fix this by always specifying the first parameter to
makeKey/makeGlobalKey as a literal string that corresponds 1:1 to
a kind of computation (usuaully specific to the getWithSet call,
or to a set of keys used in the same class/method, sans any variable
placeholders).

[1]: https://grafana.wikimedia.org/d/lqE4lcGWz/wanobjectcache-key-group

Change-Id: Id12ead6c7c56aed2be2da8e1758646c1db90393c