Remove obsolete fixme comment
In 6b0c660, the hook was updated to modern interface, which
does not provide a WikiPage anymore. The round-trip it mentions
is mostly harmless, and even though it could request a WikiPage
from WikiPageFactory, it is not guaranteed to hold a cached
revision record, in which case it is preferred to defer its
lookup for when the variable is loaded.

The remark was added in 68ff668.

Change-Id: If14afe634ca431a040fba3722775f209cdc8cec3