Load CodeEditor resources conditionally
On CI, we would like to remove the recursive injection of extensions
dependencies.  When AbuseFilter is a dependency, we are required to add
CodeEditor and WikiEditor to fullfil the requirement that all
dependencies are defined (ResourcesTest::testValidDependencies).

This makes it so that the `ext.abuseFilter.ace` resource loader module
is only registered when CodeEditor is enabled.

Bug: T389998
Change-Id: Ib295c354c90901c43a7b654fa39bb60e94bfc546
(cherry picked from commit f299b9d8042c7b53eb50346cb879b89eed745dfa)