tests: skip config reset when ConfirmEdit is not loaded
The ConfirmEditHandlerTest integration tests has a tearDown() function
invoking MediaWiki\Extension\ConfirmEdit\Hooks::getInstance(). When
ConfirmEdit is NOT loaded the test fails with:

  RuntimeException: mediaWikiSetUp() was called but not mediaWikiTearDown()
  use assertPostConditions() instead of tearDown() for post-test assertions.

The reason is the class is not available which leads to an exception and
parent::tearDown() is not invoked. In MediaWikiIntegrationTestCase,
$setupWithoutTeardown is thus not reset and the next case claims
tearDown has never been called.

This was caused by Ib223d0fee8096d1bc768bf0a49ff453384dab3ef which
rightfully moved markTestSkippedIfExtensionNotLoaded() from the test to
setUp() but missed tearDown() had a call to ConfirmEdit code.

Change-Id: I6ecb82e4c92d8100b8376c6c720da009f09e8883